+++ packages/kiarina-utils-common/src/kiarina/utils/component_registry/_types/component_factory.py
@@ -1,0 +1,5 @@
+from typing import Callable, TypeAlias, TypeVar
+
+T = TypeVar("T")
+
+ComponentFactory: TypeAlias = Callable[..., T]