+++ packages/kiarina-utils-common/src/kiarina/utils/component_registry/_types/component_input.py
@@ -1,0 +1,7 @@
+from typing import TypeAlias, TypeVar
+
+from .component_specifier import ComponentSpecifier
+
+T = TypeVar("T")
+
+ComponentInput: TypeAlias = T | ComponentSpecifier