+++ packages/kiarina-utils-common/src/kiarina/utils/component_registry/_types/component_specifier.py
@@ -1,0 +1,14 @@
+from typing import TypeAlias
+
+ComponentSpecifier: TypeAlias = str
+"""A string in one of the following formats:
+
+- {ComponentName}
+- {ComponentName}?{ConfigString}
+- {ComponentAlias}
+- {ComponentAlias}?{ConfigString}
+
+Examples:
+- my_component
+- my_component?key1=value1&key2=value2
+"""